app/api/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
tensorflow-cpu==2.8.0
numpy==1.22.0
pillow==9.0.0
gunicorn==20.1.0
h5py==3.6.0
protobuf==3.20.0
